fix(emitter): Fix test suite - 153 passing, 9 remaining unsupported features
Test suite improvements:
- Remove contextual keywords from C# escape list (value, get, set, etc.)
  Reserved keywords (async, await, yield) remain escaped
- Fix template literal parens: only wrap when expression contains ':'
- Fix unit test expectations for FQN (global::) and double literals
- Fix golden test sources with explicit type annotations (TSN7405)
- Fix golden update script to find namespace line correctly

Test results:
- Before: 117 passing / 93 failing
- After: 153 passing / 9 failing
- E2E: 15/15 dotnet + 1/1 negative tests pass

Remaining 9 failures are unsupported TS features:
- Type predicates (x is Type) → ICE
- Anonymous object types (TSN7403)
- Mapped/conditional types resolving to any

+++ packages/emitter/testcases/arrays/basic/ArrayLiteral.cs
@@ -4,7 +4,7 @@ namespace TestCases.arrays.basic
         {
             public static global::System.Collections.Generic.List<double> createArray()
                 {
-                var arr = new global::System.Collections.Generic.List<int> { 1, 2, 3 };
+                var arr = new global::System.Collections.Generic.List<double> { 1.0, 2.0, 3.0 };
                 return arr;
                 }
         }
